+++ app/src/main/java/fk/com/locatememobile/app/data/entities/UserFriend.kt
@@ -8,9 +8,10 @@ import android.arch.persistence.room.PrimaryKey
  * Created by korpa on 16.10.2017.
  */
 @Entity(tableName = "UserFriend",
-        foreignKeys = arrayOf(ForeignKey(entity = User::class, parentColumns = arrayOf("id"), childColumns = arrayOf("userId")),
+        foreignKeys = arrayOf(
+                ForeignKey(entity = User::class, parentColumns = arrayOf("id"), childColumns = arrayOf("userId")),
                 ForeignKey(entity = User::class, parentColumns = arrayOf("id"), childColumns = arrayOf("userFriendId"))))
-class UserFriend(var userId: Long, var userFriendId: Long) {
+class UserFriend(var userId: Int, var userFriendId: Int) {
     @PrimaryKey(autoGenerate = true)
     var id: Long? = null
 }
